// src/ui-client/src/models/Keyboard.ts
/* Copyright (c) 2021, UW Medicine Research IT, University of Washington
 * Developed by Nic Dobbins and Cliff Spital, CRIO Sean Mooney
 * This Source Code Form is subject to the terms of the Mozilla Public
 * License, v. 2.0. If a copy of the MPL was not distributed with this
 * file, You can obtain one at http://mozilla.org/MPL/2.0/.
 */ 

export enum keys {
    ArrowUp = 1,
    ArrowDown = 2,
    Backspace = 3,
    Enter = 4,
    Escape = 5,
    Space = 6
}